requirements.txt
pydantic-ai>=0.0.14
openai>=1.0.0
psutil>=5.9.0
python-dotenv>=1.0.0
pathlib2>=2.3.7

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
isort>=5.12.0